watch-and-learn-main/services/python-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
google-generativeai>=0.8.6
httpx>=0.27.1
pydantic>=2.7.2
python-dotenv>=1.0.0
mcp>=1.0.0
aiohttp>=3.9.0
voyageai>=0.3.0
pymongo[srv]>=4.6.0
numpy>=1.26.0
scikit-learn>=1.4.0
Pillow>=10.2.0
